backend/requirements.txt
Flask
gunicorn
opensearch-py
python-dotenv
